Change Datatype to Long at all master repositories

+++ src/main/java/com/app/smartdrive/api/repositories/master/CabrRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CabrRepository extends JpaRepository<CarBrand, Integer> {
+public interface CabrRepository extends JpaRepository<CarBrand, Long> {
 }
